Public APIs to use `Into` and `builder` function
All the public API functions were using `&str`, which is not wrong per
se, but even better approach is to use an `Into` for the public API.
Updated public APIs to use this.

Also added a `builder` function as using `default` may not be obvious,
but a documented `builder` API is a good choice.